LSNBLDR-485 Have content served through lessons wrapped.
When content is served through /access/content it can be modified on the fly be filters. One existing filter is a HTML top/tailer that adds on a header/footer to a HTML fragment. Content served through /access/lessonbuilder wasn’t getting processed in the same way.

To fix this I extracted out a service to the filter definitions are common across the whole service and exposed the collection of filters as a service so that tools outside the kernel can call it.
