+++ kernel/kernel-component/src/main/webapp/WEB-INF/content-components.xml
@@ -71,11 +71,7 @@
               <entry key="oracle" ><ref bean="org.sakaiproject.content.impl.ContentServiceSqlOracle" /></entry>
            </map>
         </property>
-        <property name="outputFilters">
-          <list>
-            <ref bean="org.sakaiproject.content.impl.HtmlPageFilter"/>
-          </list>
-        </property>
+        <property name="contentFilterService" ref="org.sakaiproject.content.api.ContentFilterService"/>
     </bean>
 
     <bean id="org.sakaiproject.content.impl.CollectionAccessFormatter"
@@ -114,6 +110,15 @@
         <property name="extensionFile"><value>content_type_extensions.properties</value></property>
         -->
 	</bean>
+
+    <bean id="org.sakaiproject.content.api.ContentFilterService"
+          class="org.sakaiproject.content.impl.ContentFilterServiceImpl">
+        <property name="outputFilters">
+            <list>
+                <ref bean="org.sakaiproject.content.impl.HtmlPageFilter"/>
+            </list>
+        </property>
+    </bean>
 
 	<bean id="org.sakaiproject.content.api.ContentResourceFilter.base"
 			class="org.sakaiproject.content.impl.BaseExtensionResourceFilter">
